requirements.txt
CTkListbox==1.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pillow==11.1.0
